testBuildings/requirements.txt
Django==1.9.6
django-debug-toolbar==1.4
djangorestframework==3.3.3
sqlparse==0.1.19
